added home component for render in register.js
