+++ client/src/pages/styles/login.css
@@ -1,5 +1,5 @@
 body, html{
-    background-image: url("D:\10X Asignmnts\Car-Rental-Project\car-rental-app\src\pages\styles\images\car-rental-background.png");
+    background-image: url("D:\10X Asignmnts\Car-Rental-Project\car-rental-app\client\src\pages\styles\images\car-rental-background.png");
     background-repeat: no-repeat;
     background-size: cover;
     min-height: 100%;
